statistics: move to lockstats package
