+++ statistics/handle/BUILD.bazel
@@ -11,7 +11,6 @@ go_library(
         "handle_hist.go",
         "historical_stats_handler.go",
         "lock_stats_handler.go",
-        "unlock_stats_handler.go",
         "update.go",
     ],
     importpath = "github.com/pingcap/tidb/statistics/handle",
@@ -32,6 +31,7 @@ go_library(
         "//sessiontxn",
         "//statistics",
         "//statistics/handle/cache",
+        "//statistics/handle/lockstats",
         "//statistics/handle/metrics",
         "//table",
         "//types",
@@ -61,14 +61,13 @@ go_test(
         "dump_test.go",
         "gc_test.go",
         "handle_hist_test.go",
-        "lock_stats_handler_test.go",
         "main_test.go",
         "update_list_test.go",
     ],
     embed = [":handle"],
     flaky = True,
     race = "on",
-    shard_count = 29,
+    shard_count = 26,
     deps = [
         "//config",
         "//domain",
